Remove legacy gql types and use fragment approach for complex request shape
